# raido-openapi-3.0.yaml
openapi: 3.0.3
servers:
  - url: 'https://api.prod.raid.org.au'
    description: The production service - "PROD".
  - url: 'https://api.demo.raid.org.au'
    description: >
      DEMO environment, used for client integration testing and 
      customer demonstrations.
  - url: 'https://api.test.raid.org.au'
    description: >
      dev environment, used for developer integration testing.
  - url: 'http://localhost:8080'
    description: Local development.
info:
  title: RAID v2 API
  version: 2.0.0
  description: >
    This file is where all the endpoint paths are defined, it's the "top level'
    of the OpenAPI definition that links all the different files together.
    The `3.0` in the filename refers to this file being based on OpenAPI 3.0 
    as opposed to OpenAPI 3.1, which the tooling doesn't support yet.
    The `2.0.0` in the version field refers to the fact that there's already 
    a `1.0.0` used for the legacy RAiD application.
    Note that swagger ui doesn't currently work with our spec, 
    see https://github.com/swagger-api/swagger-ui/issues/7724
    But the spec works fine with openapi-generator tooling.

  contact:
    name: ARDC RAiD team
    email: contact@raid.org
  termsOfService: https://github.com/au-research/raido/blob/main/doc/service-level-guide.md
...
security:
  - bearerAuth: []
paths:
...
  /raid/:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: mintRaid
      summary: Mint a raid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaidCreateRequest'
      responses:
        201:
          description: The Raid was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaidDto'
        400:
          description: There was a problem with the message sent to the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailureResponse'
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: findAllRaids
      summary: List raids
      parameters:
        - name: includeFields
          description: The top level fields to include in each RAiD in the response body. Excludes all other fields.
          example: identifier,title,date
          in: query
          schema:
            type: array
            items:
              type: string
        - name: contributor.id
          description: Only show RAiDs that include a contributor with the given id
          example: https://orcid.org/0009-0006-4129-5257
          in: query
          schema:
            type: string
        - name: organisation.id
          description: Only show RAiDs that include an organisation with the given id
          example: https://ror.org/038sjwq14
          in: query
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaidDto'
  /raid/{prefix}/{suffix}:
    parameters:
      - name: prefix
        in: path
        required: true
        description: Prefix of the raid handle
        schema:
          type: string
      - name: suffix
        in: path
        required: true
        description: Suffix of the raid handle
        schema:
          type: string
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: findRaidByName
      summary: Read a raid
      responses:
        200:
          description: data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaidDto'
        403:
          description: Closed or Embargoed raids return a 403 response with an access statement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClosedRaid'
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: updateRaid
      summary: Update a raid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaidUpdateRequest'
      responses:
        200:
          description: The Raid was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaidDto'
        400:
          description: Bad request. There was a problem with the message sent to the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailureResponse'
    patch:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: patchRaid
      summary: Patch a raid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaidPatchRequest'
      responses:
        200:
          description: The Raid was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RaidDto'
        400:
          description: Bad request. There was a problem with the message sent to the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailureResponse'
  /raid/{prefix}/{suffix}/{version}:
    parameters:
      - name: prefix
        in: path
        required: true
        description: Prefix of the raid handle
        schema:
          type: string
      - name: suffix
        in: path
        required: true
        description: Suffix of the raid handle
        schema:
          type: string
      - name: version
        in: path
        required: true
        description: The version of the raid
        schema:
          type: integer
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: findRaidByNameAndVersion
      summary: Read a raid with a specified version
      responses:
        200:
          description: data
          content:
            application/json:
              schema:
                type: object
        403:
          description: Closed or Embargoed raids return a 403 response with an access statement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClosedRaid'
...
  /raid/all-public:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - raid
      operationId: findAllPublicRaids
      summary: List raids
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaidDto'
  /service-point/:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - service-point
      operationId: createServicePoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePointCreateRequest'
      responses:
        201:
          description: The service point was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePoint'
        400:
          description: There was a problem with the message sent to the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailureResponse'
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - service-point
      operationId: findAllServicePoints
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicePoint'
  /service-point/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The id of the service point
        schema:
          type: integer
          format: int64
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - service-point
      operationId: findServicePointById
      responses:
        200:
          description: data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePoint'
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - service-point
      operationId: updateServicePoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePointUpdateRequest'
      responses:
        200:
          description: The service point was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePoint'
        400:
          description: Bad request. There was a problem with the message sent to the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailureResponse'

  # endregion
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RaidCreateRequest:
      type: object
      required:
        - metadataSchema
        - titles
        - dates
        - contributors
        - access
      properties:
        metadata:
          $ref: '#/components/schemas/Metadata'
        identifier:
          $ref: '#/components/schemas/Id'
        title:
          type: array
          items: { $ref: '#/components/schemas/Title' }
        date: { $ref: '#/components/schemas/Date' }
        description:
          type: array
          items: { $ref: '#/components/schemas/Description' }
        access: { $ref: '#/components/schemas/Access' }
        alternateUrl:
          type: array
          items: { $ref: '#/components/schemas/AlternateUrl' }
        contributor:
          type: array
          items: { $ref: '#/components/schemas/Contributor' }
        organisation:
          type: array
          items: { $ref: '#/components/schemas/Organisation' }
        subject:
          type: array
          items: { $ref: '#/components/schemas/Subject' }
        relatedRaid:
          type: array
          items: { $ref: '#/components/schemas/RelatedRaid' }
        relatedObject:
          type: array
          items: { $ref: '#/components/schemas/RelatedObject' }
        alternateIdentifier:
          type: array
          items: { $ref: '#/components/schemas/AlternateIdentifier' }
        spatialCoverage:
          type: array
          items: { $ref: '#/components/schemas/SpatialCoverage' }
    RaidUpdateRequest:
      type: object
      required:
        - identifier
        - metadataSchema
        - titles
        - dates
        - contributors
        - access
      properties:
        metadata: { $ref: '#/components/schemas/Metadata' }
        identifier: { $ref: '#/components/schemas/Id' }
        title:
          type: array
          items: { $ref: '#/components/schemas/Title' }
        date: { $ref: '#/components/schemas/Date' }
        description:
          type: array
          items: { $ref: '#/components/schemas/Description' }
        access: { $ref: '#/components/schemas/Access' }
        alternateUrl:
          type: array
          items: { $ref: '#/components/schemas/AlternateUrl' }
        contributor:
          type: array
          items: { $ref: '#/components/schemas/Contributor' }
        organisation:
          type: array
          items: { $ref: '#/components/schemas/Organisation' }
        subject:
          type: array
          items: { $ref: '#/components/schemas/Subject' }
        relatedRaid:
          type: array
          items: { $ref: '#/components/schemas/RelatedRaid' }
        relatedObject:
          type: array
          items: { $ref: '#/components/schemas/RelatedObject' }
        alternateIdentifier:
          type: array
          items: { $ref: '#/components/schemas/AlternateIdentifier' }
        spatialCoverage:
          type: array
          items: { $ref: '#/components/schemas/SpatialCoverage' }
    RaidPatchRequest:
      type: object
      properties:
        contributor:
          type: array
          items: { $ref: '#/components/schemas/Contributor'}
    RaidDto:
      allOf:
        - $ref: '#/components/schemas/RaidUpdateRequest'
      type: object
      properties:
        traditionalKnowledgeLabel:
          type: array
          items: { $ref: '#/components/schemas/TraditionalKnowledgeLabel' }
    Metadata:
      type: object
      properties:
        created:
          type: number
          format: datetime
        updated:
          type: number
          format: datetime
    Id:
      type: object
      required: [ id, schemaUri, registrationAgency, owner, servicePoint, license, version ]
      properties:
        id:
          type: string
          description: The identifier of the raid, e.g. https://raid.org.au/102.100.100/zzz
          example: https://raid.org/xxx.yyy/zzz
        schemaUri:
          type: string
          description: >
            The URI of the Identifier scheme. For example, https://raid.org
          example: https://raid.org
        registrationAgency:
          $ref: '#/components/schemas/RegistrationAgency'
        owner:
          $ref: '#/components/schemas/Owner'
        raidAgencyUrl:
          type: string
          description: >
            The URL for the raid via the minting raid agency system
          example: https://static.raid.org.au/raid/10378.1/1695863
        license:
          type: string
          description: The license under which the RAiD Metadata Record associated with this Identifier has been issued.
          example: Creative Commons CC-0
        version:
          type: integer
          description: The version of the resource. Read-only. Increments automatically on update.
    RegistrationAgency:
      type: object
      description: >
        ROR that identifies the organisation that operates the raid registration agency software that minted this raid.
      required: [id, schemaUri]
      properties:
        id: {
          type: string,
          example: https://ror.org/02stey378
        }
        schemaUri: {
          type: string,
          example: https://ror.org
        }
    Owner:
      type: object
      description: >
        The legal entity responsible for the RAiD; the ‘Owner’ of a RAiD. Analogous to a DataCite ‘Member’, has a 
        legal agreement with the Registration Agency.
      required: [id, schemaUri]
      properties:
        id: {
          type: string,
          example: https://ror.org/02stey378
        }
        schemaUri: {
          type: string,
          example: https://ror.org
        }
        servicePoint:
          type: integer
          format: int64
          description: >
            The Service Point (SP) that requested the RAiD. Analogous to a DataCite ‘Repository’. SPs belong to an owner, RAiD owners can have multiple SPs, and SPs do not need to be legal entities.
          example: 20000003
    Date:
      type: object
      description: Metadata schema block containing the start and end date of the RAiD.
      required: [ "startDate" ]
      properties:
        startDate: {
          type: string,
          example: 2023-08-28; 2023-08; 2023
        }
        endDate: {
          type: string,
          example: 2023-08-28; 2023-08; 2023
        }
    Title:
      type: object
      required: [text, type, startDate]
      properties:
        text: {type: string}
        type:
          $ref: '#/components/schemas/TitleType'
        startDate: {
          type: string,
          example: 2023-08-28; 2023-08; 2023
        }
        endDate: {
          type: string,
          example: 2023-08-28; 2023-08; 2023
        }
        language:
          $ref: '#/components/schemas/Language'
    TitleType:
      type: object
      properties:
        id: {
          type: string,
          example: https://vocabulary.raid.org/title.type.schema/5
        }
        schemaUri: {
          type: string,
          example: https://vocabulary.raid.org/title.type.schema/376
        }
    Description:
      type: object
      required: ["text", "type"]
      properties:
        text: {type: string}
        type:
          $ref: '#/components/schemas/DescriptionType'
        language:
          $ref: '#/components/schemas/Language'
    DescriptionType:
      type: object
      required: [id, schemaUri]
      properties:
        id: {
          type: string,
          example: https://vocabulary.raid.org/description.type.schema/318
        }
        schemaUri: {
          type: string,
          example: https://vocabulary.raid.org/description.type.schema/320
        }
    Access:
      type: object
      required: ["type"]
      properties:
        type:
          $ref: '#/components/schemas/AccessType'
        statement:
          $ref: '#/components/schemas/AccessStatement'
        embargoExpiry: {
          type: string,
          format: date,
          description: 'Date the embargo on access to the RAiD ends. Year, month, and day required; may not be more than 18 months from the date the RAiD was registered.'
        }
    AccessStatement:
      type: object
      required: [statement]
      properties:
        text: { type: string }
        language:
          $ref: '#/components/schemas/Language'
    AccessType:
      type: object
      required: [ id, schemaUri ]
      properties:
        id: {
          type: string,
          example: https://vocabularies.coar-repositories.org/access_rights/c_abf2/,
          description: Open access
        }
        schemaUri: {
          type: string,
          example: https://vocabularies.coar-repositories.org/access_rights/
        }
    Contributor:
      type: object
      required: [id, schemaUri, position, role]
      properties:
        id:
          type: string
        schemaUri: {
          type: string,
          example: https://orcid.org/
        }
        status:
          description: 'Read only. Whether the contributor has confirmed they are a part of the project. One of PENDING_AUTHENTICATION, AUTHENTICATED, UNAUTHENTICATED, FAILED'
          type: string
        statusMessage:
          description: 'Read only. If there is an error when attempting to verify the contributor the failure message will appear here'
          type: string
        email:
          type: string
        uuid:
          type: string
        position:
          type: array
          items:
            $ref: '#/components/schemas/ContributorPosition'
        role:
          type: array
          items:
            $ref: '#/components/schemas/ContributorRole'
        leader:
          type: boolean
        contact:
          type: boolean
    ContributorPosition:
      type: object
      required: [ schemaUri, id, startDate ]
      properties:
        schemaUri: {
          type: string,
          example: https://vocabulary.raid.org/contributor.position.schema/305
        }
        id: {
          type: string,
          example: https://vocabulary.raid.org/contributor.position.schema/307,
          description: Principal or Lead Investigator
        }
        startDate: {
          type: string,
          example: 2025-08-28; 2025-08; 2025
        }
        endDate: {
          type: string,
          example: 2025-08-28; 2025-08; 2025
        }
    ContributorRole:
      type: object
      required: [ id, schemaUri ]
      properties:
        schemaUri: {
          type: string,
          example: https://credit.niso.org/
        }
        id: {
          type: string,
          example: https://credit.niso.org/contributor-roles/writing-review-editing/
        }
    Organisation:
      type: object
      required: [id, schemaUri, role]
      properties:
        id: { type: string }
        schemaUri: {
          type: string,
          example: https://ror.org/
        }
        role:
          type: array
          items:
            $ref: '#/components/schemas/OrganisationRole'
    OrganisationRole:
      type: object
      required: [id, schemaUri, startDate]
      properties:
        schemaUri: {
          type: string,
          example: https://vocabulary.raid.org/organisation.role.schema/359
        }
        id: {
          type: string,
          example: https://vocabulary.raid.org/organisation.role.schema/182,
          description: Lead Research Organisation
        }
        startDate: {
          type: string,
          example: 2025-08-28; 2025-08; 2025
        }
        endDate: {
          type: string,
          example: 2025-08-28; 2025-08; 2025
        }
    AlternateUrl:
      type: object
      required: [url]
      properties:
        url: { type: string }
    Subject:
      type: object
      required: [ id, schemaUri ]
      properties:
        id: { type: string }
        schemaUri: { type: string }
        keyword:
          type: array
          items:
            $ref: '#/components/schemas/SubjectKeyword'
...
    RelatedRaid:
      type: object
      properties:
        id: { type: string }
        type:
          $ref: '#/components/schemas/RelatedRaidType'
...
    RelatedObject:
      type: object
      properties:
        id: { type: string }
        schemaUri: { type: string }
        type:
          $ref: '#/components/schemas/RelatedObjectType'
        category:
          type: array
          items:
            $ref: '#/components/schemas/RelatedObjectCategory'
...
    AlternateIdentifier:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
    SpatialCoverage:
      type: object
      properties:
        id: { type: string }
        schemaUri: { type: string }
        place:
          type: array
          items:
            $ref: '#/components/schemas/SpatialCoveragePlace'
...
    TraditionalKnowledgeLabel:
      type: object
      properties:
        id: {
          type: string,
          example: https://localcontexts.org/label/tk-verified/
        }
        schemaUri: {
          type: string,
          example: https://localcontexts.org/labels/traditional-knowledge-labels/
        }
    Language:
      type: object
      required: [ code, schemaUri]
      properties:
        id: {
          type: string,
          example: eng
        }
        schemaUri: {
          type: string,
          example: https://iso639-3.sil.org/
        }
    ClosedRaid:
      type: object
      properties:
        identifier:
          $ref: '#/components/schemas/Id'
        access:
          $ref: '#/components/schemas/Access'
...
    ServicePointCreateRequest:
      type: object
      required:
        - name
        - identifierOwner
        - groupId
      properties:
        name:
          type: string
          description: The name of the service point
          example: Australian Research Data Commons
        adminEmail:
          type: string
          description: The email address of the person responsible for administering the service point.
        techEmail:
          type: string
          description: The email address of a technical contact when using the API
        identifierOwner:
          type: string
          description: The PID of the institution that will own the RAiDs (currently only RORs are supported).
          example: https://ror.org/038sjwq14
        repositoryId:
          type: string
          description: The Datacite repository id. This needs to be created in Fabrica.
          example: VBGR.QJYTRX
        groupId:
          type: string
          description: The Keycloak group id (UUID) associated with the service point.
          example: ba0b01a6-726f-464f-b501-454a10096826
        prefix:
          type: string
          description: The prefix used in the handle when minting RAiDs. Assigned when the repository is created.
          example: 10.82481
        password:
          type: string
          description: The password of the repository in Datacite.
        appWritesEnabled:
          type: boolean
          description: Whether users are able to edit RAiDs in the in the app. This can cause conflicts when also creating/updating RAiDs through the API.
        enabled:
          type: boolean
          description: Whether the service point is able to create or update RAiDs
    ServicePointUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/ServicePointCreateRequest'
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          description: The name of the service point
          example: Australian Research Data Commons
    ServicePoint:
      description: The response for all service point requests
      type: object
      required: [ id, name, identifierOwner, adminEmail, techEmail, enabled ]
      properties:
        id: { $ref: '#/components/schemas/SurrogateKey' }
        name: { type: string }
        identifierOwner: { type: string }
        repositoryId:
          type: string
          description: The Datacite repository id. This needs to be created in Fabrica.
          example: VBGR.QJYTRX
        prefix:
          type: string
          description: The prefix used in the handle when minting RAiDs. Assigned when the repository is created.
          example: 10.82481
        groupId:
          type: string
          description: The Keycloak group id (UUID) associated with the service point.
          example: ba0b01a6-726f-464f-b501-454a10096826
        searchContent: { type: string }
        techEmail: { type: string }
        adminEmail: { type: string }
        enabled: { type: boolean }
        appWritesEnabled: { type: boolean }
    SurrogateKey:
      type: integer
      format: int64
    ValidationFailure:
      type: object
      required: [ fieldId, errorType, message ]
      properties:
        fieldId: { type: string }
        errorType: { type: string }
        message: { type: string }
    FailureResponse:
      type: object
      required: [ type, title, status, detail, instance ]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
    ValidationFailureResponse:
      allOf:
        - $ref: '#/components/schemas/FailureResponse'
      required: [ failures ]
      properties:
        failures:
          type: array
          items:
            $ref: '#/components/schemas/ValidationFailure'